Added cart service

+++ src/main/java/com/bush/pharmacy_web_app/controllers/http/CatalogController.java
@@ -33,7 +33,7 @@ public class CatalogController {
 
     @GetMapping("/{id}")
     public String getProduct(Model model,
-                             @PathVariable Integer id) {
+                             @PathVariable Long id) {
         var product = service.findById(id).orElseThrow();
 
         model.addAttribute("product", product);
